+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next version
 - Improve FTP097 to also consider `StrEnum` and `IntEnum`
+- FTP030 will report an import of `from __future__ import annotations` if python 3.14 onwards is used
 
 ## 25.5.22.0
 - Support async for loops for all checks working on for loops